# swagger-mongodb/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Backend Alternatives
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /moment:
    x-swagger-router-controller: moment_service
    # Collection - GET
    get:
      description: Returns a specific moment to the caller
      operationId: getMoments
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Moment"
        "400":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    # Collection - POST
    post:
      description: Creates a new moment
      operationId: postMoment
      parameters:
        - name: moment
          in: body
          schema:
            $ref: "#/definitions/Moment"
          description: The new moment to be posted
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Moment"
        "400":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /moment/{id}:
    x-swagger-router-controller: moment_service
    # Element - GET
    get:
      description: Retrieve a specific moment
      operationId: getMoment
      parameters:
        - name: id
          in: path
          description: The id of the moment to retrieve
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Moment"
        "400":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    # Element - PUT
    put:
      description: Modifies a given moment according to the submitted object
      operationId: putMoment
      parameters:
        - name: id
          in: path
          description: The id of the moment to modify
          required: true
          type: string
        - name: changes
          in: body
          schema:
            $ref: "#/definitions/Moment"
          description: A moment object describing the changes
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Moment"
        "400":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    # Element - DELETE
    delete:
      description: Delete a specific moment
      operationId: deleteMoment
      parameters:
        - name: id
          in: path
          description: The id of the moment to modify
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Moment"
        "400":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /moment/{id}/comment:
    x-swagger-router-controller: moment_service
    # Collection - GET (query)
    get:
      description: Returns list of all comments which have been posted to the given moment
      operationId: getComments
      parameters:
        - name: id
          in: path
          description: The id of the parent moment
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Comment"
        "400":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    # Collection - POST
    post:
      description: Posts a new comment to the given moment
      operationId: postComment
      parameters:
        - name: id
          in: path
          description: The id of the parent moment
          required: true
          type: string
        - name: comment
          in: body
          schema:
            $ref: "#/definitions/Comment"
          description: The new comment to be posted
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Comment"
        "400":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /moment/{id}/comment/{cid}:
    x-swagger-router-controller: moment_service
    # Element - DELETE
    delete:
      description: Delete a specific comment from the given moment
      operationId: deleteComment
      parameters:
        - name: id
          in: path
          description: The id of the parent moment
          required: true
          type: string
        - name: cid
          in: path
          description: The id of the comment to delete
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Comment"
        "400":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Moment:
    required:
      - authorName
      - text
    properties:
      _id:
        type: string
      authorName:
        type: string
      text:
        type: string
        maxLength: 300
      created:
        type: string
      modified:
        type: string
      commentCount:
        type: number
      __v:
        type: number
  Comment:
    required:
      - authorName
      - text
      - parentMoment
    properties:
      _id:
        type: string
      authorName:
        type: string
      parent:
        type: string
      parentMoment:
        type: string
      text:
        type: string
        maxLength: 300
      created:
        type: string
      modified:
        type: string
      __v:
        type: number
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string